심심해서 해본 텍스트파일 읽어서 객체 만든 후 전체 출력하기

+++ Lecture/Date221006/ReadFileExample/Student.java
@@ -3,9 +3,9 @@ package Lecture.Date221006.ReadFileExample;
 public class Student {
     private int id;
     private String hid;
-    private int height;
+    private double height;
 
-    public Student(int id, String hid, int height) {
+    public Student(int id, String hid, double height) {
         this.id = id;
         this.hid = hid;
         this.height = height;
